chore(session) do not read body by default and only on certain HTTP methods
### Summary

On comment: https://github.com/Kong/kong/issues/7418#issuecomment-1411729346
@PidgeyBE mentioned that session plugin reads bodies by on every HTTP request
that is not an GET request.

Because it is quite common to use bodies to send large files, reading body
makes features like route.request_buffering=off, not working. Thus, the default
value for ´logout_post_arg` in session plugin was removed. The bodies are only
read when this is configured. This might change behavior on scripts that create
session plugin and which also think that logout by body argument works as before.
On the other hand it is a way more common to read session than it is to log out
session, thus it should be better value for future of us.